♻️ 1F9PPF task: split PR check helpers
